Restructuring file with new names
