+++ .eslintrc.json
@@ -1,0 +1,4 @@
+{
+    "root": true,
+    "extends": "@quoin/eslint-config-quoin"
+}
