+++ lib/HSRuntime.js
@@ -3,7 +3,7 @@ var fs = require('fs');
 var path = require('path');
 
 const MonAppError = require('./error');
-const config = require('./../config');
+const config = require('./config');
 
 //
 // Class "HSRuntime"
