+++ lib/index.js
@@ -1,0 +1,5 @@
+const app = require('./app');
+
+module.exports = {
+    app
+};
