+++ lib/routes/apiRoutes.js
@@ -2,8 +2,8 @@ var express = require('express');
 var appApiRouter = express.Router();
 var ObjectID = require("mongodb").ObjectID;
 
-var $rt = require('./../src/HSRuntime.js');
-const MonAppError = require('./../src/error');
+var $rt = require('./../HSRuntime.js');
+const MonAppError = require('./../error');
 
 appApiRouter.post('/CRUD', function(req, res, next) {
     if (req.xhr || req.accepts('json,html') === 'json') {
@@ -148,7 +148,7 @@ appApiRouter.post('/CRUD', function(req, res, next) {
                         }
 
                         console.log("ID1:" + entity._id);
-                        var idQuery = {"_id": new ObjectID(entity._id)};
+                        var idQuery = {_id: new ObjectID(entity._id)};
                         delete entity._id;
                         console.log("Collection: " + collection.collectionName);
