+++ public/app/appUtilities.js
