--- runtime/.eslintrc.json
@@ -1,24 +1,0 @@
-{
-    "root": true,
-    "extends": "standard",
-    "plugins": [
-        "standard",
-        "promise"
-    ],
-    "rules": {
-        "brace-style": ["error", "1tbs", {
-            "allowSingleLine": false
-        }],
-        "curly": ["error", "all"],
-        "indent": ["error", 4, {
-            "SwitchCase": 1
-        }],
-        "quotes": "off",
-        "semi": ["error", "always"],
-        "space-before-function-paren": ["error", {
-            "anonymous": "never",
-            "named": "never",
-            "asyncArrow": "ignore"
-        }]
-    }
-}
