--- runtime/.eslintrc.test.json
@@ -1,6 +1,0 @@
-{
-  "extends": "./.eslintrc.json",
-  "env": {
-    "mocha": true
-  }
-}
