--- runtime/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "MonApp",
-  "version": "0.6.1",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/dslama/MonApp"
-  },
-  "engines": {
-    "node": ">=6.9.0"
-  },
-  "scripts": {
-    "eslint": "eslint --fix --ignore-pattern *.test.js server",
-    "start": "node ./bin/www"
-  },
-  "dependencies": {
-    "body-parser": "~1.15.1",
-    "cookie-parser": "~1.4.3",
-    "debug": "~2.2.0",
-    "express": "~4.13.4",
-    "hbs": "~4.0.0",
-    "hbs-utils": "0.0.4",
-    "mongodb": "2.2.22",
-    "morgan": "~1.7.0",
-    "rc": "1.1.7",
-    "serve-favicon": "~2.3.0"
-  },
-  "devDependencies": {
-    "eslint": "3.18.0",
-    "eslint-config-standard": "7.1.0",
-    "eslint-plugin-promise": "3.5.0",
-    "eslint-plugin-standard": "2.1.1"
-  }
-}
