+++ server.js
@@ -7,14 +7,14 @@
 var debug = require('debug')('SimpleSite:server');
 var http = require('http');
 
-var app = require('../app');
-const config = require('./../server/config');
+var app = require('./lib').app;
+const config = require('./lib/config');
 
 /**
  * Get port from environment and store in Express.
  */
 
-var port = normalizePort(config.port);
+var port = normalizePort(process.env.PORT || config.port || 3001);
 app.set('port', port);
 
 /**
@@ -36,19 +36,19 @@ server.on('listening', onListening);
  */
 
 function normalizePort(val) {
-  var port = parseInt(val, 10);
+    var port = parseInt(val, 10);
 
-  if (isNaN(port)) {
+    if (isNaN(port)) {
     // named pipe
-    return val;
-  }
+        return val;
+    }
 
-  if (port >= 0) {
+    if (port >= 0) {
     // port number
-    return port;
-  }
+        return port;
+    }
 
-  return false;
+    return false;
 }
 
 /**
@@ -56,27 +56,27 @@ function normalizePort(val) {
  */
 
 function onError(error) {
-  if (error.syscall !== 'listen') {
-    throw error;
-  }
+    if (error.syscall !== 'listen') {
+        throw error;
+    }
 
-  var bind = typeof port === 'string'
+    var bind = typeof port === 'string'
     ? 'Pipe ' + port
     : 'Port ' + port;
 
   // handle specific listen errors with friendly messages
-  switch (error.code) {
-    case 'EACCES':
-      console.error(bind + ' requires elevated privileges');
-      process.exit(1);
-      break;
-    case 'EADDRINUSE':
-      console.error(bind + ' is already in use');
-      process.exit(1);
-      break;
-    default:
-      throw error;
-  }
+    switch (error.code) {
+        case 'EACCES':
+            console.error(bind + ' requires elevated privileges');
+            process.exit(1);
+            break;
+        case 'EADDRINUSE':
+            console.error(bind + ' is already in use');
+            process.exit(1);
+            break;
+        default:
+            throw error;
+    }
 }
 
 /**
@@ -84,9 +84,9 @@ function onError(error) {
  */
 
 function onListening() {
-  var addr = server.address();
-  var bind = typeof addr === 'string'
+    var addr = server.address();
+    var bind = typeof addr === 'string'
     ? 'pipe ' + addr
     : 'port ' + addr.port;
-  debug('Listening on ' + bind);
+    debug('Listening on ' + bind);
 }
